Updated token to make it optional.
please test this on the JsonReq function, i have tested this on XHR (client side). if there is no token passed, the header is not set. 
in the server side, you still cannot have a null value / undefined value in the header. (therefore the ternary operator logic doesn't work)
ie, "Authorization : bearer {null} here"
so there is a if else check to build the Jsonreq, if token is present the header parameter is set. otherwise it's not. 
i assume that if user has a invalid token we'd like to show them that

feel free to request additional changes.